streamlit_app/requirements.txt
streamlit==1.26.0
pandas
sqlalchemy
psycopg2-binary
plotly
